Added TicketService class
